fix: time units are not normalized (issue #122) (#318)
This PR fixes an issue when the new benchmark results are using a different time unit than the previous one. This happens specifically when the benchmark values are very close to the boundary between 2 units, ex. `990 ns/iter` and `1.1 us/iter`. In that case, we need to normalize the newly coming results to the previously used unit so that the values are comparable.

* extract addBenchmarkEntry function
* normalize new entry values to match last entry units
* handle `<time>`, `<time>/iter` and `ops/<time>` units